TP4 - long long for perfs

+++ TP4/lib_matrix.c
@@ -18,7 +18,7 @@ void dgetf2(struct matrix* A) {
     }
 }
 
-int dgetf2_complexity(int size) {
+long long dgetf2_complexity(int size) {
     return size * size * (size - 1)/2;
 }
 
@@ -216,7 +216,7 @@ void matrix_solve(struct matrix* A, struct vector* X, struct vector* B) {
     }
 }
 
-int matrix_solve_complexity(int size) {
+long long matrix_solve_complexity(int size) {
     return 2*size*(size-1) + 3*size;
 }
